Refactor factory-related types and enhance API functionality
Consolidate factory-related types by moving definitions from forge.ts to factory.ts, improving organization and clarity. Update various services and models to utilize the new factory types, ensuring consistent usage across the codebase. Refactor app generation logic to await asynchronous operations, enhancing reliability. Additionally, introduce new error handling for blockchain interactions, improving overall system robustness.
