+++ src/types/index.ts
@@ -1,3 +1,3 @@
 export * from './db.ts'
-export * from './forge.ts'
+export * from './factory.ts'
 export * from './llm.ts'
